+++ packages/zowe-explorer/__tests__/__unit__/trees/shared/SharedUtils.unit.test.ts
@@ -12,7 +12,7 @@
 import * as vscode from "vscode";
 import { createIProfile, createISession, createInstanceOfProfile } from "../../../__mocks__/mockCreators/shared";
 import { createDatasetSessionNode } from "../../../__mocks__/mockCreators/datasets";
-import { createUSSNode } from "../../../__mocks__/mockCreators/uss";
+import { createUSSNode, createUSSSessionNode } from "../../../__mocks__/mockCreators/uss";
 import { UssFSProvider } from "../../../../src/trees/uss/UssFSProvider";
 import { imperative, ProfilesCache, Gui, ZosEncoding, BaseProvider } from "@zowe/zowe-explorer-api";
 import { Constants } from "../../../../src/configuration/Constants";
@@ -21,12 +21,14 @@ import { ZoweLocalStorage } from "../../../../src/tools/ZoweLocalStorage";
 import { ZoweLogger } from "../../../../src/tools/ZoweLogger";
 import { DatasetFSProvider } from "../../../../src/trees/dataset/DatasetFSProvider";
 import { ZoweDatasetNode } from "../../../../src/trees/dataset/ZoweDatasetNode";
-import { ZoweJobNode } from "../../../../src/trees/job/ZoweJobNode";
+import { ZoweJobNode, ZoweSpoolNode } from "../../../../src/trees/job/ZoweJobNode";
 import { SharedUtils } from "../../../../src/trees/shared/SharedUtils";
 import { ZoweUSSNode } from "../../../../src/trees/uss/ZoweUSSNode";
 import { AuthUtils } from "../../../../src/utils/AuthUtils";
 import { SharedTreeProviders } from "../../../../src/trees/shared/SharedTreeProviders";
 import { MockedProperty } from "../../../__mocks__/mockUtils";
+import { createIJobFile, createJobSessionNode } from "../../../__mocks__/mockCreators/jobs";
+import { JobFSProvider } from "../../../../src/trees/job/JobFSProvider";
 
 function createGlobalMocks() {
     const newMocks = {
@@ -507,7 +509,7 @@ describe("Shared utils unit tests - function promptForEncoding", () => {
             profile: blockMocks.profile,
             parentNode: sessionNode,
         });
-        DatasetFSProvider.instance.encodingMap["TEST.PS"] = { kind: "text" };
+        DatasetFSProvider.instance.encodingMap[node.resourceUri?.path] = { kind: "text" };
         blockMocks.getEncodingForFile.mockReturnValueOnce(undefined);
         await SharedUtils.promptForEncoding(node);
         expect(blockMocks.showQuickPick).toHaveBeenCalled();
@@ -535,6 +537,512 @@ describe("Shared utils unit tests - function promptForEncoding", () => {
         await SharedUtils.promptForEncoding(node);
         expect(blockMocks.showQuickPick).toHaveBeenCalled();
         expect(blockMocks.showQuickPick.mock.calls[0][1]).toEqual(expect.objectContaining({ placeHolder: "Current encoding is IBM-1047" }));
+    });
+
+    it("remembers cached encoding for spool node", async () => {
+        const blockMocks = createBlockMocks();
+        const sessionNode = createJobSessionNode(blockMocks.session, blockMocks.profile);
+        const jobNode = new ZoweJobNode({
+            label: "TESTPS",
+            collapsibleState: vscode.TreeItemCollapsibleState.Collapsed,
+            session: blockMocks.session,
+            profile: blockMocks.profile,
+            parentNode: sessionNode,
+        });
+        const spoolNode = new ZoweSpoolNode({
+            label: "SPOOL",
+            collapsibleState: vscode.TreeItemCollapsibleState.None,
+            spool: createIJobFile(),
+            parentNode: jobNode,
+        });
+        JobFSProvider.instance.encodingMap[spoolNode.resourceUri?.path] = { kind: "text" };
+        blockMocks.getEncodingForFile.mockReturnValueOnce(undefined);
+        await SharedUtils.promptForEncoding(spoolNode);
+        expect(blockMocks.showQuickPick).toHaveBeenCalled();
+        expect(blockMocks.showQuickPick.mock.calls[0][1]).toEqual(expect.objectContaining({ placeHolder: "Current encoding is EBCDIC" }));
+    });
+
+    it("prompts for text encoding for Spool file", async () => {
+        const blockMocks = createBlockMocks();
+        const sessionNode = createJobSessionNode(blockMocks.session, blockMocks.profile);
+        const jobNode = new ZoweJobNode({
+            label: "TESTPS",
+            collapsibleState: vscode.TreeItemCollapsibleState.Collapsed,
+            session: blockMocks.session,
+            profile: blockMocks.profile,
+            parentNode: sessionNode,
+        });
+        const node = new ZoweSpoolNode({
+            label: "testFile",
+            collapsibleState: vscode.TreeItemCollapsibleState.None,
+            session: blockMocks.session,
+            profile: blockMocks.profile,
+            parentNode: jobNode,
+            spool: createIJobFile(),
+        });
+
+        blockMocks.showQuickPick.mockImplementationOnce(async (items) => items[0]);
+        blockMocks.getEncodingForFile.mockReturnValueOnce(undefined);
+        const encoding = await SharedUtils.promptForEncoding(node);
+        expect(blockMocks.showQuickPick).toHaveBeenCalled();
+        expect(encoding).toEqual(textEncoding);
+    });
+
+    it("prompts for binary encoding for Spool file", async () => {
+        const blockMocks = createBlockMocks();
+        const sessionNode = createJobSessionNode(blockMocks.session, blockMocks.profile);
+        const jobNode = new ZoweJobNode({
+            label: "TESTPS",
+            collapsibleState: vscode.TreeItemCollapsibleState.Collapsed,
+            session: blockMocks.session,
+            profile: blockMocks.profile,
+            parentNode: sessionNode,
+        });
+        const node = new ZoweSpoolNode({
+            label: "testFile",
+            collapsibleState: vscode.TreeItemCollapsibleState.None,
+            session: blockMocks.session,
+            profile: blockMocks.profile,
+            parentNode: jobNode,
+            spool: createIJobFile(),
+        });
+
+        blockMocks.showQuickPick.mockImplementationOnce(async (items) => items[1]);
+        const encoding = await SharedUtils.promptForEncoding(node);
+        expect(blockMocks.showQuickPick).toHaveBeenCalled();
+        expect(encoding).toEqual(binaryEncoding);
+    });
+
+    it("prompts for other encoding for Spool file and returns codepage", async () => {
+        const blockMocks = createBlockMocks();
+        const sessionNode = createJobSessionNode(blockMocks.session, blockMocks.profile);
+        const jobNode = new ZoweJobNode({
+            label: "TESTPS",
+            collapsibleState: vscode.TreeItemCollapsibleState.Collapsed,
+            session: blockMocks.session,
+            profile: blockMocks.profile,
+            parentNode: sessionNode,
+        });
+        const node = new ZoweSpoolNode({
+            label: "testFile",
+            collapsibleState: vscode.TreeItemCollapsibleState.None,
+            session: blockMocks.session,
+            profile: blockMocks.profile,
+            parentNode: jobNode,
+            spool: createIJobFile(),
+        });
+
+        blockMocks.showQuickPick.mockImplementationOnce(async (items) => items[2]);
+        blockMocks.showInputBox.mockResolvedValueOnce("IBM-1047");
+        const encoding = await SharedUtils.promptForEncoding(node);
+        expect(blockMocks.showQuickPick).toHaveBeenCalled();
+        expect(blockMocks.showInputBox).toHaveBeenCalled();
+        expect(encoding).toEqual(otherEncoding);
+    });
+
+    it("prompts for other encoding for Spool file and returns undefined", async () => {
+        const blockMocks = createBlockMocks();
+        const sessionNode = createJobSessionNode(blockMocks.session, blockMocks.profile);
+        const jobNode = new ZoweJobNode({
+            label: "TESTPS",
+            collapsibleState: vscode.TreeItemCollapsibleState.Collapsed,
+            session: blockMocks.session,
+            profile: blockMocks.profile,
+            parentNode: sessionNode,
+        });
+        const node = new ZoweSpoolNode({
+            label: "testFile",
+            collapsibleState: vscode.TreeItemCollapsibleState.None,
+            session: blockMocks.session,
+            profile: blockMocks.profile,
+            parentNode: jobNode,
+            spool: createIJobFile(),
+        });
+
+        blockMocks.showQuickPick.mockImplementationOnce(async (items) => items[2]);
+        blockMocks.showInputBox.mockResolvedValueOnce(undefined);
+        const encoding = await SharedUtils.promptForEncoding(node);
+        expect(blockMocks.showQuickPick).toHaveBeenCalled();
+        expect(blockMocks.showInputBox).toHaveBeenCalled();
+        expect(encoding).toBeUndefined();
+    });
+});
+
+describe("Shared utils unit tests - function getCachedEncoding", () => {
+    const mockSession = createISession();
+    const mockProfile = createIProfile();
+    describe("Spool nodes", () => {
+        it("correctly returns the cached encoding for binary", async () => {
+            const sessionNode = createJobSessionNode(mockSession, mockProfile);
+            const jobNode = new ZoweJobNode({
+                label: "TESTPS",
+                collapsibleState: vscode.TreeItemCollapsibleState.Collapsed,
+                session: mockSession,
+                profile: mockProfile,
+                parentNode: sessionNode,
+            });
+            const node = new ZoweSpoolNode({
+                label: "testFile",
+                collapsibleState: vscode.TreeItemCollapsibleState.None,
+                session: mockSession,
+                profile: mockProfile,
+                parentNode: jobNode,
+                spool: createIJobFile(),
+            });
+            const encoding = { kind: "binary" };
+
+            const encodingMapSpy = jest.spyOn(node, "getEncodingInMap").mockResolvedValue(encoding);
+            const response = await SharedUtils.getCachedEncoding(node);
+
+            expect(encodingMapSpy).toHaveBeenCalledWith(node.resourceUri?.path);
+            expect(response).toEqual(encoding.kind);
+        });
+
+        it("correctly returns the cached encoding for text", async () => {
+            const sessionNode = createJobSessionNode(mockSession, mockProfile);
+            const jobNode = new ZoweJobNode({
+                label: "TESTPS",
+                collapsibleState: vscode.TreeItemCollapsibleState.Collapsed,
+                session: mockSession,
+                profile: mockProfile,
+                parentNode: sessionNode,
+            });
+            const node = new ZoweSpoolNode({
+                label: "testFile",
+                collapsibleState: vscode.TreeItemCollapsibleState.None,
+                session: mockSession,
+                profile: mockProfile,
+                parentNode: jobNode,
+                spool: createIJobFile(),
+            });
+            const encoding = { kind: "text" };
+
+            const encodingMapSpy = jest.spyOn(node, "getEncodingInMap").mockResolvedValue(encoding);
+            const response = await SharedUtils.getCachedEncoding(node);
+
+            expect(encodingMapSpy).toHaveBeenCalledWith(node.resourceUri?.path);
+            expect(response).toEqual(encoding.kind);
+        });
+
+        it("correctly returns the cached encoding for other", async () => {
+            const sessionNode = createJobSessionNode(mockSession, mockProfile);
+            const jobNode = new ZoweJobNode({
+                label: "TESTPS",
+                collapsibleState: vscode.TreeItemCollapsibleState.Collapsed,
+                session: mockSession,
+                profile: mockProfile,
+                parentNode: sessionNode,
+            });
+            const node = new ZoweSpoolNode({
+                label: "testFile",
+                collapsibleState: vscode.TreeItemCollapsibleState.None,
+                session: mockSession,
+                profile: mockProfile,
+                parentNode: jobNode,
+                spool: createIJobFile(),
+            });
+            const encoding = { kind: "other", codepage: "IBM-1147" };
+
+            const encodingMapSpy = jest.spyOn(node, "getEncodingInMap").mockResolvedValue(encoding);
+            const response = await SharedUtils.getCachedEncoding(node);
+
+            expect(encodingMapSpy).toHaveBeenCalledWith(node.resourceUri?.path);
+            expect(response).toEqual(encoding.codepage);
+        });
+
+        it("correctly returns the cached encoding for undefined", async () => {
+            const sessionNode = createJobSessionNode(mockSession, mockProfile);
+            const jobNode = new ZoweJobNode({
+                label: "TESTPS",
+                collapsibleState: vscode.TreeItemCollapsibleState.Collapsed,
+                session: mockSession,
+                profile: mockProfile,
+                parentNode: sessionNode,
+            });
+            const node = new ZoweSpoolNode({
+                label: "testFile",
+                collapsibleState: vscode.TreeItemCollapsibleState.None,
+                session: mockSession,
+                profile: mockProfile,
+                parentNode: jobNode,
+                spool: createIJobFile(),
+            });
+            const encoding = undefined;
+
+            const encodingMapSpy = jest.spyOn(node, "getEncodingInMap").mockResolvedValue(encoding);
+            const response = await SharedUtils.getCachedEncoding(node);
+
+            expect(encodingMapSpy).toHaveBeenCalledWith(node.resourceUri?.path);
+            expect(response).toEqual(encoding);
+        });
+    });
+
+    describe("USS nodes", () => {
+        it("correctly returns the cached encoding for binary", async () => {
+            const sessionNode = createUSSSessionNode(mockSession, mockProfile);
+            const ussNode = new ZoweUSSNode({
+                label: "TEST.PS",
+                collapsibleState: vscode.TreeItemCollapsibleState.None,
+                contextOverride: Constants.USS_BINARY_FILE_CONTEXT,
+                session: mockSession,
+                profile: mockProfile,
+                parentNode: sessionNode,
+            });
+            const encoding = { kind: "binary" };
+
+            const encodingMapSpy = jest.spyOn(ussNode, "getEncodingInMap").mockResolvedValue(encoding);
+            const response = await SharedUtils.getCachedEncoding(ussNode);
+
+            expect(encodingMapSpy).toHaveBeenCalledWith(ussNode.resourceUri?.path);
+            expect(response).toEqual(encoding.kind);
+        });
+
+        it("correctly returns the cached encoding for text", async () => {
+            const sessionNode = createUSSSessionNode(mockSession, mockProfile);
+            const ussNode = new ZoweUSSNode({
+                label: "TEST.PS",
+                collapsibleState: vscode.TreeItemCollapsibleState.None,
+                contextOverride: Constants.USS_TEXT_FILE_CONTEXT,
+                session: mockSession,
+                profile: mockProfile,
+                parentNode: sessionNode,
+            });
+            const encoding = { kind: "text" };
+
+            const encodingMapSpy = jest.spyOn(ussNode, "getEncodingInMap").mockResolvedValue(encoding);
+            const response = await SharedUtils.getCachedEncoding(ussNode);
+
+            expect(encodingMapSpy).toHaveBeenCalledWith(ussNode.resourceUri?.path);
+            expect(response).toEqual(encoding.kind);
+        });
+
+        it("correctly returns the cached encoding for other", async () => {
+            const sessionNode = createUSSSessionNode(mockSession, mockProfile);
+            const ussNode = new ZoweUSSNode({
+                label: "TEST.PS",
+                collapsibleState: vscode.TreeItemCollapsibleState.None,
+                contextOverride: Constants.USS_TEXT_FILE_CONTEXT,
+                session: mockSession,
+                profile: mockProfile,
+                parentNode: sessionNode,
+            });
+            const encoding = { kind: "other", codepage: "IBM-1147" };
+
+            const encodingMapSpy = jest.spyOn(ussNode, "getEncodingInMap").mockResolvedValue(encoding);
+            const response = await SharedUtils.getCachedEncoding(ussNode);
+
+            expect(encodingMapSpy).toHaveBeenCalledWith(ussNode.resourceUri?.path);
+            expect(response).toEqual(encoding.codepage);
+        });
+
+        it("correctly returns the cached encoding for undefined", async () => {
+            const sessionNode = createUSSSessionNode(mockSession, mockProfile);
+            const ussNode = new ZoweUSSNode({
+                label: "TEST.PS",
+                collapsibleState: vscode.TreeItemCollapsibleState.None,
+                contextOverride: Constants.USS_TEXT_FILE_CONTEXT,
+                session: mockSession,
+                profile: mockProfile,
+                parentNode: sessionNode,
+            });
+            const encoding = undefined;
+
+            const encodingMapSpy = jest.spyOn(ussNode, "getEncodingInMap").mockResolvedValue(encoding);
+            const response = await SharedUtils.getCachedEncoding(ussNode);
+
+            expect(encodingMapSpy).toHaveBeenCalledWith(ussNode.resourceUri?.path);
+            expect(response).toEqual(encoding);
+        });
+    });
+
+    describe("Dataset nodes", () => {
+        describe("Sequential", () => {
+            it("correctly returns the cached encoding for binary", async () => {
+                const sessionNode = createDatasetSessionNode(mockSession, mockProfile);
+                const dsNode = new ZoweDatasetNode({
+                    label: "TEST.PS",
+                    collapsibleState: vscode.TreeItemCollapsibleState.None,
+                    contextOverride: Constants.DS_DS_BINARY_CONTEXT,
+                    session: mockSession,
+                    profile: mockProfile,
+                    parentNode: sessionNode,
+                });
+                const encoding = { kind: "binary" };
+
+                const encodingMapSpy = jest.spyOn(dsNode, "getEncodingInMap").mockResolvedValue(encoding);
+                const response = await SharedUtils.getCachedEncoding(dsNode);
+
+                expect(encodingMapSpy).toHaveBeenCalledWith(dsNode.resourceUri?.path);
+                expect(response).toEqual(encoding.kind);
+            });
+
+            it("correctly returns the cached encoding for text", async () => {
+                const sessionNode = createDatasetSessionNode(mockSession, mockProfile);
+                const dsNode = new ZoweDatasetNode({
+                    label: "TEST.PS",
+                    collapsibleState: vscode.TreeItemCollapsibleState.None,
+                    contextOverride: Constants.DS_DS_CONTEXT,
+                    session: mockSession,
+                    profile: mockProfile,
+                    parentNode: sessionNode,
+                });
+                const encoding = { kind: "test" };
+
+                const encodingMapSpy = jest.spyOn(dsNode, "getEncodingInMap").mockResolvedValue(encoding);
+                const response = await SharedUtils.getCachedEncoding(dsNode);
+
+                expect(encodingMapSpy).toHaveBeenCalledWith(dsNode.resourceUri?.path);
+                expect(response).toEqual(encoding.kind);
+            });
+
+            it("correctly returns the cached encoding for other", async () => {
+                const sessionNode = createDatasetSessionNode(mockSession, mockProfile);
+                const dsNode = new ZoweDatasetNode({
+                    label: "TEST.PS",
+                    collapsibleState: vscode.TreeItemCollapsibleState.None,
+                    contextOverride: Constants.DS_DS_CONTEXT,
+                    session: mockSession,
+                    profile: mockProfile,
+                    parentNode: sessionNode,
+                });
+                const encoding = { kind: "other", codepage: "IBM-1147" };
+
+                const encodingMapSpy = jest.spyOn(dsNode, "getEncodingInMap").mockResolvedValue(encoding);
+                const response = await SharedUtils.getCachedEncoding(dsNode);
+
+                expect(encodingMapSpy).toHaveBeenCalledWith(dsNode.resourceUri?.path);
+                expect(response).toEqual(encoding.codepage);
+            });
+
+            it("correctly returns the cached encoding for undefined", async () => {
+                const sessionNode = createDatasetSessionNode(mockSession, mockProfile);
+                const dsNode = new ZoweDatasetNode({
+                    label: "TEST.PS",
+                    collapsibleState: vscode.TreeItemCollapsibleState.None,
+                    contextOverride: Constants.DS_DS_CONTEXT,
+                    session: mockSession,
+                    profile: mockProfile,
+                    parentNode: sessionNode,
+                });
+                const encoding = undefined;
+
+                const encodingMapSpy = jest.spyOn(dsNode, "getEncodingInMap").mockResolvedValue(encoding);
+                const response = await SharedUtils.getCachedEncoding(dsNode);
+
+                expect(encodingMapSpy).toHaveBeenCalledWith(dsNode.resourceUri?.path);
+                expect(response).toEqual(encoding);
+            });
+        });
+
+        describe("Partitioned", () => {
+            it("correctly returns the cached encoding for binary", async () => {
+                const sessionNode = createDatasetSessionNode(mockSession, mockProfile);
+                const dsNode = new ZoweDatasetNode({
+                    label: "TEST.PDS",
+                    collapsibleState: vscode.TreeItemCollapsibleState.Collapsed,
+                    contextOverride: Constants.DS_PDS_CONTEXT,
+                    session: mockSession,
+                    profile: mockProfile,
+                    parentNode: sessionNode,
+                });
+                const memNode = new ZoweDatasetNode({
+                    label: "TESTMEM",
+                    collapsibleState: vscode.TreeItemCollapsibleState.None,
+                    contextOverride: Constants.DS_MEMBER_BINARY_CONTEXT,
+                    session: mockSession,
+                    profile: mockProfile,
+                    parentNode: dsNode,
+                });
+                const encoding = { kind: "binary" };
+
+                const encodingMapSpy = jest.spyOn(memNode, "getEncodingInMap").mockResolvedValue(encoding);
+                const response = await SharedUtils.getCachedEncoding(memNode);
+
+                expect(encodingMapSpy).toHaveBeenCalledWith(memNode.resourceUri?.path);
+                expect(response).toEqual(encoding.kind);
+            });
+
+            it("correctly returns the cached encoding for text", async () => {
+                const sessionNode = createDatasetSessionNode(mockSession, mockProfile);
+                const dsNode = new ZoweDatasetNode({
+                    label: "TEST.PDS",
+                    collapsibleState: vscode.TreeItemCollapsibleState.Collapsed,
+                    contextOverride: Constants.DS_PDS_CONTEXT,
+                    session: mockSession,
+                    profile: mockProfile,
+                    parentNode: sessionNode,
+                });
+                const memNode = new ZoweDatasetNode({
+                    label: "TESTMEM",
+                    collapsibleState: vscode.TreeItemCollapsibleState.None,
+                    contextOverride: Constants.DS_MEMBER_CONTEXT,
+                    session: mockSession,
+                    profile: mockProfile,
+                    parentNode: dsNode,
+                });
+                const encoding = { kind: "text" };
+
+                const encodingMapSpy = jest.spyOn(memNode, "getEncodingInMap").mockResolvedValue(encoding);
+                const response = await SharedUtils.getCachedEncoding(memNode);
+
+                expect(encodingMapSpy).toHaveBeenCalledWith(memNode.resourceUri?.path);
+                expect(response).toEqual(encoding.kind);
+            });
+
+            it("correctly returns the cached encoding for other", async () => {
+                const sessionNode = createDatasetSessionNode(mockSession, mockProfile);
+                const dsNode = new ZoweDatasetNode({
+                    label: "TEST.PDS",
+                    collapsibleState: vscode.TreeItemCollapsibleState.Collapsed,
+                    contextOverride: Constants.DS_PDS_CONTEXT,
+                    session: mockSession,
+                    profile: mockProfile,
+                    parentNode: sessionNode,
+                });
+                const memNode = new ZoweDatasetNode({
+                    label: "TESTMEM",
+                    collapsibleState: vscode.TreeItemCollapsibleState.None,
+                    contextOverride: Constants.DS_MEMBER_CONTEXT,
+                    session: mockSession,
+                    profile: mockProfile,
+                    parentNode: dsNode,
+                });
+                const encoding = { kind: "other", codepage: "IBM-1147" };
+
+                const encodingMapSpy = jest.spyOn(memNode, "getEncodingInMap").mockResolvedValue(encoding);
+                const response = await SharedUtils.getCachedEncoding(memNode);
+
+                expect(encodingMapSpy).toHaveBeenCalledWith(memNode.resourceUri?.path);
+                expect(response).toEqual(encoding.codepage);
+            });
+
+            it("correctly returns the cached encoding for undefined", async () => {
+                const sessionNode = createDatasetSessionNode(mockSession, mockProfile);
+                const dsNode = new ZoweDatasetNode({
+                    label: "TEST.PDS",
+                    collapsibleState: vscode.TreeItemCollapsibleState.Collapsed,
+                    contextOverride: Constants.DS_PDS_CONTEXT,
+                    session: mockSession,
+                    profile: mockProfile,
+                    parentNode: sessionNode,
+                });
+                const memNode = new ZoweDatasetNode({
+                    label: "TESTMEM",
+                    collapsibleState: vscode.TreeItemCollapsibleState.None,
+                    contextOverride: Constants.DS_MEMBER_CONTEXT,
+                    session: mockSession,
+                    profile: mockProfile,
+                    parentNode: dsNode,
+                });
+                const encoding = undefined;
+
+                const encodingMapSpy = jest.spyOn(memNode, "getEncodingInMap").mockResolvedValue(encoding);
+                const response = await SharedUtils.getCachedEncoding(memNode);
+
+                expect(encodingMapSpy).toHaveBeenCalledWith(memNode.resourceUri?.path);
+                expect(response).toEqual(encoding);
+            });
+        });
     });
 });