+++ packages/zowe-explorer/src/trees/shared/SharedUtils.ts
@@ -143,11 +143,11 @@ export class SharedUtils {
     public static async getCachedEncoding(node: IZoweTreeNode): Promise<string | undefined> {
         let cachedEncoding: ZosEncoding;
         if (SharedUtils.isZoweUSSTreeNode(node)) {
-            cachedEncoding = await node.getEncodingInMap(node.fullPath);
+            cachedEncoding = await node.getEncodingInMap(node.resourceUri.path);
+        } else if (SharedUtils.isZoweJobTreeNode(node)) {
+            cachedEncoding = await node.getEncodingInMap(node.resourceUri.path);
         } else {
-            const isMemberNode = node.contextValue.startsWith(Constants.DS_MEMBER_CONTEXT);
-            const dsKey = isMemberNode ? `${node.getParent().label as string}(${node.label as string})` : (node.label as string);
-            cachedEncoding = await (node as IZoweDatasetTreeNode).getEncodingInMap(dsKey);
+            cachedEncoding = await (node as IZoweDatasetTreeNode).getEncodingInMap(node.resourceUri.path);
         }
         return cachedEncoding?.kind === "other" ? cachedEncoding.codepage : cachedEncoding?.kind;
     }
@@ -190,7 +190,10 @@ export class SharedUtils {
             .filter(Boolean);
     }
 
-    public static async promptForEncoding(node: IZoweDatasetTreeNode | IZoweUSSTreeNode, taggedEncoding?: string): Promise<ZosEncoding | undefined> {
+    public static async promptForEncoding(
+        node: IZoweDatasetTreeNode | IZoweUSSTreeNode | IZoweJobTreeNode,
+        taggedEncoding?: string
+    ): Promise<ZosEncoding | undefined> {
         const ebcdicItem: vscode.QuickPickItem = {
             label: vscode.l10n.t("EBCDIC"),
             description: vscode.l10n.t("z/OS default codepage"),